add create project functions

+++ products/models.py
@@ -20,7 +20,7 @@ class Category(models.Model):
 class Product(models.Model):
     name = models.CharField(max_length=100)
     description = models.TextField()
-    price = models.DecimalField(max_digits=5, decimal_places=2)
+    price = models.PositiveIntegerField()
     product_image = models.ImageField(upload_to='static/images/products/')
     category = models.ForeignKey(Category, on_delete=models.CASCADE)
     slug = models.SlugField(max_length=100, unique=True, null=True, blank=True)
